+++ pepe-tg/src/plugins/fakeRaresPlugin.ts
@@ -20,6 +20,7 @@ import { CardFactsImportService } from '../services/CardFactsImportService';
 import { NewCardService } from '../services/NewCardService';
 import { ReminderService } from '../services/ReminderService';
 import { BacklogService } from '../services/BacklogService';
+import { ArtistSpotlightService } from '../services/ArtistSpotlightService';
 import { noteScrillaMention } from '../conversation/anniversaryRuntime';
 import { noteShown, recentlyShown } from '../utils/cardShowCooldown';
 import { findRepeat } from '../utils/repeatGuard';
@@ -873,7 +874,7 @@ export const fakeRaresPlugin: Plugin = {
   // gates were built to close. If it is ever wanted, route it through
   // gateSubmission first.
   evaluators: [],
-  services: [KnowledgeOrchestratorService, MemoryStorageService, TelemetryService, CardDisplayService, SmartRouterService, XHarvestService, RecapService, SocialMemoryService, ReleaseNoteService, AnniversaryService, CardFactsImportService, NewCardService, ReminderService, BacklogService],
+  services: [KnowledgeOrchestratorService, MemoryStorageService, TelemetryService, CardDisplayService, SmartRouterService, XHarvestService, RecapService, SocialMemoryService, ReleaseNoteService, AnniversaryService, CardFactsImportService, NewCardService, ReminderService, BacklogService, ArtistSpotlightService],
   
   events: {
     MESSAGE_RECEIVED: [